+++ backend/scheduler/run.sh
@@ -7,3 +7,4 @@ python src/scheduler.py
 
 
 
+