+++ mobile/lib/widgets/token_limit_banner.dart
@@ -103,3 +103,4 @@ class TokenLimitBanner extends StatelessWidget {
   }
 }
 
+