+++ setup.ps1
@@ -59,3 +59,4 @@ Write-Host "3. Flutter 앱을 실행하세요" -ForegroundColor White
 
 
 
+